+++ packages/shared/src/server/index.ts
@@ -11,6 +11,8 @@ export * from "./services/PromptService/types";
 export * from "./services/traces-ui-table-service";
 export * from "./services/InMemoryFilterService";
 export * from "./services/DatasetService";
+export * from "./datasets/schemaValidation";
+export * from "./datasets/schemaTypes";
 export * from "./evalJobConfigCache";
 export * from "./auth/apiKeys";
 export * from "./auth/invalidateApiKeys";
