+++ packages/shared/src/server/services/DatasetService/index.ts
@@ -1,4 +1,3 @@
 export * from "./DatasetItemManager";
-export * from "./utils/schemaValidation";
 export * from "./types";
 export * from "./DatasetItemValidator";
