+++ packages/shared/src/server/services/DatasetService/types.ts
@@ -1,38 +1,5 @@
-import { z } from "zod/v4";
-import { ValidationError } from "ajv";
-import { Dataset, DatasetStatus, Prisma } from "../../../db";
-import { jsonSchemaNullable } from "../../../utils/zod";
-import {
-  FieldValidationError,
-  isValidJSONSchema,
-} from "../../../utils/jsonSchemaValidation";
-
-/**
- * Schema for validating JSON Schema objects
- * Supports JSON Schema Draft 2020-12
- * Uses Ajv to validate that the input is a valid JSON Schema
- */
-export const DatasetJSONSchema = z
-  .record(z.string(), jsonSchemaNullable)
-  .refine(isValidJSONSchema, {
-    message: "Must be a valid JSON Schema",
-  });
-
-export type DatasetJSONSchema = z.infer<typeof DatasetJSONSchema>;
-
-/**
- * Result type for dataset mutations that may fail validation
- * Using discriminated union for type-safe error handling
- */
-export type DatasetMutationResult =
-  | {
-      success: true;
-      dataset: Dataset;
-    }
-  | {
-      success: false;
-      validationErrors: ValidationError[];
-    };
+import { DatasetStatus, Prisma } from "../../../db";
+import { FieldValidationError } from "../../../utils/jsonSchemaValidation";
 
 export type PayloadError = {
   success: false;
