+++ web/src/features/datasets/server/dataset-router.ts
@@ -9,7 +9,6 @@ import { auditLog } from "@/src/features/audit-logs/auditLog";
 import { DB } from "@/src/server/db";
 import {
   paginationZod,
-  DatasetStatus,
   singleFilter,
   StringNoHTML,
   StringNoHTMLNonEmpty,
@@ -48,72 +47,15 @@ import {
   getDatasetItemIdsWithRunData,
   DatasetItemManager,
   validateAllDatasetItems,
-  validateDatasetItemField,
-  DatasetItemValidator,
   DatasetJSONSchema,
   type DatasetMutationResult,
 } from "@langfuse/shared/src/server";
-import { createId as createCuid } from "@paralleldrive/cuid2";
 import { aggregateScores } from "@/src/features/scores/lib/aggregateScores";
 import {
   updateDataset,
   upsertDataset,
 } from "@/src/features/datasets/server/actions/createDataset";
 import { type BulkDatasetItemValidationError } from "@langfuse/shared";
-
-/**
- * Remove problematic C0 and C1 control characters from string values.
- * PostgreSQL TEXT columns cannot store NULL byte (\u0000) and other control characters.
- * Preserves common characters like newlines and tabs.
- */
-const cleanControlChars = (input: string): string => {
-  if (!input) return input;
-
-  // Remove control characters:
-  // \u0000-\u0008: NULL through backspace
-  // \u000B: vertical tab (preserve \n=\u000A, \t=\u0009, \r=\u000D)
-  // \u000E-\u001F: shift out through unit separator
-  // \u007F-\u009F: DEL + C1 controls
-  return input.replace(/[\u0000-\u0008\u000B\u000E-\u001F\u007F-\u009F]/g, "");
-};
-
-/**
- * Recursively clean control characters from all string values in a JSON structure.
- * This handles strings within objects and arrays after JSON.parse.
- */
-const sanitizeJsonValue = (value: unknown): unknown => {
-  if (typeof value === "string") {
-    return cleanControlChars(value);
-  }
-  if (Array.isArray(value)) {
-    return value.map(sanitizeJsonValue);
-  }
-  if (value && typeof value === "object") {
-    return Object.fromEntries(
-      Object.entries(value).map(([k, v]) => [k, sanitizeJsonValue(v)]),
-    );
-  }
-  return value;
-};
-
-const formatDatasetItemData = (data: string | null | undefined) => {
-  if (data === "") return Prisma.DbNull;
-
-  try {
-    const parsed = !!data ? JSON.parse(data) : undefined;
-    // Sanitize control characters from parsed object before sending to PostgreSQL
-    return parsed
-      ? (sanitizeJsonValue(parsed) as Prisma.InputJsonObject)
-      : undefined;
-  } catch (e) {
-    logger.info(
-      "[trpc.datasets.formatDatasetItemData] failed to parse dataset item data",
-      e,
-    );
-
-    return undefined;
-  }
-};
 
 /**
  * Adds a case-insensitive search condition to a Kysely query
@@ -773,10 +715,10 @@ export const datasetRouter = createTRPCRouter({
         resourceType: "datasetItem",
         resourceId: input.datasetItemId,
         action: "update",
-        after: result.updatedItem,
+        after: result.datasetItem,
       });
 
-      return result.updatedItem;
+      return result.datasetItem;
     }),
   createDataset: protectedProjectProcedure
     .input(
